Remove support for reading YAML files
We already introduced deprecation warnings back in
49b889b03eb5e39f4fc348966a989567d58, telling people that we're going to
remove YAML in version 3.0.

With this, the removal is essentially done and we now no longer depend
on yaml-cpp, nor do we need pkg-config or anything like that, because we
essentially have zero (external) library dependencies except libc and
libdl.

This of course also greatly simplifies our parsing code and also gets
rid of a few tests that now became obsolete.

Signed-off-by: aszlig <aszlig@nix.build>