Remove leftover references to YAML from usage help
This is a leftover which I forgot to remove when deprecating support for
YAML files. Besides getting that stuff out of the documentation in
general to discourage use, the description was even wrong and
misleading, since it referred to RULES_FILE and RULES_DATA - both of
which no longer exist since they got removed in 49b889b03eb5e39f4fc3489.

Signed-off-by: aszlig <aszlig@nix.build>