README: Document what "unknown" means in %p/%a/%t
First of all, the port number could *in* *theory* also end up with the
value "unknown", so I documented it accordingly as well.

Second, the reason I mention "in theory" is because we currently only
pick up sockets that are either AF_INET or AF_INET6, but due to a bug or
in a future release where we support more socket types something like
this could happen.

One of these is already something I do have in mind and already prepared
for it (see globpath.cc) is support for matching on Unix domain sockets,
because while we're already pretty good at bending IP sockets to our
will, but we really can't do anything if the application in question
already uses AF_UNIX (or AF_LOCAL) sockets.

The only thing that could really happen right now is if the socket type
is something like SOCK_PACKET, SOCK_RAW, SOCK_RDM or SOCK_SEQPACKET.

Signed-off-by: aszlig <aszlig@nix.build>