README: Fix program usage synopsis
This is solely for the manpage and with the deprecation of the YAML file
option, we also have less restrictive ways of specifying rules and we
can now mix rule files and inline rule arguments as we please.

While this was fixed in the usage of the program itself, the manpage
still had the old synopsis in it, which with this commit now no longer
is the case.

Signed-off-by: aszlig <aszlig@nix.build>