README: Break out build instructions into INSTALL
I've choosen the name INSTALL, since this is how autotools usually calls
the file containing build instructions.

The reason for moving it to a separate file is because the README is
already getting quite long and since the build instructions are only
relevant for people actually wanting to use ip2unix, it doesn't need to
be in the README, nor do they need to be in the manpage.

Signed-off-by: aszlig <aszlig@nix.build>