usage: Fall back to URL instead of manpage
If we don't build the manpage (eg. if AsciiDoc or Asciidoctor is
unavailable) it doesn't really make a lot of sense if we refer to the
manpage from the command line usage.

To avoid users from being confused about the missing manpage, let's at
least provide an URL to the rule specification section of the README.

Signed-off-by: aszlig <aszlig@nix.build>