Integrate LGTM code analysis
Unfortunately, autodetected by LGTM didn't work quite well and it only
detected the project as a Python project since the tests are in Python.

It seems that LGTM is using some kind of compiler drop-in that connects
to some remote host, at least it looks like it given the following error
message:

  [build] cc1plus: error: /usr/local/include/x86_64-linux-gnu: Connection refused

So instead let's add a configuration file giving more hints to LGTM on
how to analyse the project.

I've tested this using their "test analysis configuration" feature and
if there is no other magic involved this new configuration should work.

Also, I'm not quite sure whether the path_classifiers to explicitly
specify the tests subdirectory is necessary, but given that we *only*
have tests in that directory it probably makes sense to be explicit
about it.

Signed-off-by: aszlig <aszlig@nix.build>