lgtm.yml: Bring in par with working test config
Comparing the current .lgtm.yml with the working test configuration I
submitted manually yields a small difference:

>>> yaml.load(open('.lgtm.yml', 'r'))
{'path_classifiers': {'test': [{'exclude': '/'}, 'tests']},
 'extraction': {'cpp': {'configure': {'command': ['meson setup build']},
                        'index': {'build_command': ['cd build', 'ninja']}}}}
>>> yaml.load(open('working.yml', 'r'))
{'path_classifiers': {'test': [{'exclude': '/'}, 'tests']},
 'extraction': {'cpp': {'configure': {'command': 'meson setup build'},
                        'index': {'build_command': ['cd build', 'ninja']}}}}

The difference here is that the configure command is a list in our file
but is a plain string in the working test config.

Additionally I also added quotes around strings to make sure that we
don't risk any weird magic that might implicitly recognise those strings
as a different type.

Signed-off-by: aszlig <aszlig@nix.build>