README: Add badge for LGTM
Right now the destination URL still runs into an error page, but I think
this is because LGTM didn't yet update the project statistics.

Neveretheless, there currently is one alert about the FIXME comment we
have in sockaddr.cc, which indicates that the analysis was done.

The reason for adding the badge is mainly to annoy me so that once my
OCD kicks in, I'll fix the alerts in order to get to an A+ score :-)

Signed-off-by: aszlig <aszlig@nix.build>