preload: Fix incrementing rule position
Regression introduced by 1a8a9b6ca73865a10452eb10c93dd751d0ba6af7.

The commit in question overhauled the way we do rule matching for
systemd sockets, but when changing the match_rule function to return the
rule position in the parent commit (1761a000f5159a8fef449eabf1e1d10e84),
I forgot to actually increment the variable.

For normal actions, the rule position doesn't matte much because we
immediately perform the action in case of a match. For systemd rules on
the other hand, we need to know whether we need the rule position
matches with the corresponding systemd socket.

This and the fact that we only had tests for single systemd sockets was
the reason why we did not catch this earlier. To avoid future
regressions, I added a small NixOS VM test with multiple socket fds.

Signed-off-by: aszlig <aszlig@nix.build>