flake: Fix evaluation of rsession test
Evaluation of that test returns the following error:

  access to URI '.../flake-compat/...' is forbidden in restricted mode

The reason is that whenever we call rsession.nix without an ip2unix
attribute, it defaults to being imported from the top-level default.nix.

However, with Flakes, the default.nix is now a compatibility expression
and uses fetchTarball, is not allowed from within Flakes. So let's
actually remove the default args and pass ip2unix from the Flake so that
we now no longer should be running any impure functions.

After fixing this however, we got another evaluation error:

  Package ‘hunspell-dict-nl-nl-2.20.19’ in ... has an unfree license

This is because RStudio depends on Hunspell and the latter apparently
was marked as unfree in recent nixpkgs versions.

For the test, we don't need Hunspell dictionaries, so I simply added an
override to disable them.

Signed-off-by: aszlig <aszlig@nix.build>