sockaddr: Turn FIXME comment into XXX
The LGTM service constantly annoys me with this alert, which in fact
isn't actually a thing that's broken, because the function in question
is just slower than it could be.

This is going to be addressed at some point[1] but it's not something
critical and since "FIXME" is the wrong label anyway, let's get rid of
that annoying alert.

[1]: https://github.com/nixcloud/ip2unix/issues/25

Signed-off-by: aszlig <aszlig@nix.build>