flake: Only use a subset of systems for Hydra
While it's certainly okay to allow to build the Flake package for all
supported systems, my own Hydra however currently has only support for
building i686-linux and x86_64-linux.

So to make sure we are not riddled with build errors, let's introduce an
attribute that specifies just the systems supported by Hydra.

Signed-off-by: aszlig <aszlig@nix.build>