tests/vm: Remove cruft from debugging
Of course we don't want to sleep for 20 seconds if a test command fails,
but I had this in during debugging because when tracing the ip2unix call
the test already stopped while journal logs were still in the pipeline,
thus truncating the logs.

In the systemd-single test, I just had a leftover print-like statement.

Signed-off-by: aszlig <aszlig@nix.build>