tests: Add a Meson dependency on the main binary
The release notes[1] for Meson 0.57 state:

  Until now, meson test rebuilt the whole project independent of the
  requested tests and their dependencies. With this release, meson test
  will only rebuild what is needed for the tests or suites that will be
  run. This feature can be used, for example, to speed up bisecting
  regressions using commands like the following:

  git bisect start <broken commit> <working commit>
  git bisect run meson test <failing test name>

  This would find the broken commit automatically while at each step
  rebuilding only those pieces of code needed to run the test.

  However, this change could cause failures when upgrading to 0.57, if
  the dependencies are not specified correctly in meson.build.

We're passing the main binary to pytest, but it seems that Meson doesn't
recognise this as a dependency if we're using ip2unix.full_path().

With our Nix build, we don't have this problem because we're always
building the full project and run the tests afterwards, but the GitHub
Actions workflow fails due to that because it is only running
"meson test" without prior building of the main target.

Adding an explicit dependency to the integration test function fixes
this.

[1]: https://mesonbuild.com/Release-notes-for-0-57-0.html#meson-test-only-rebuilds-test-dependencies

Signed-off-by: aszlig <aszlig@nix.build>