Use different internal name for preloaded symbols
When building the project with CLang, we get the following compile
error:

  ../src/preload.cc:349:29: error: redefinition of 'recvfrom'
  extern "C" ssize_t WRAP_SYM(recvfrom)(int fd, void *buf, size_t len, int flags,
                              ^
  ...-glibc-2.32-48-dev/include/bits/socket2.h:64:1: note: previous definition is here
  recvfrom (int __fd, void *__restrict __buf, size_t __n, int __flags,
  ^

The quick (and dirty) fix for this would be by using yet another
"#if{,def}" and be more or less done, but as soon as we port this to
more platforms, we're certainly going to accumulate a long list of those
conditionals.

However, a better solution is to not actually re-define those functions
with the same name but use an internal version with a prefix and alias
it to the correct symbol name when linking.

This is exactly what we're doing now and I tested this with both GCC and
CLang and it's running fine with all the tests.

Signed-off-by: aszlig <aszlig@nix.build>