scripts/gensyms: Use real C++ preprocessor
So far we've skipped all the preprocessor directives and just used a
regular expression to check whether one of WRAP_SYM or EXPORT_SYM
occurs in one of the source files.

While this works fine if we're using a linker version script, it fails
when using PROVIDE-directives because the symbols actually need to
exist.

Some of our wrapped functions however are dependent on the
configuration, for example we only wrap the listen function if we have
systemd support enabled.

To fix this, we're now using the real preprocessor on all of the source
files while also ignoring "#include" directives. Of course, ignoring
them is also a bit limited, especially if we relocate the *_SYM macros,
but for very simple definitions like we have how it's sufficient.

This should fix building without systemd support.

Signed-off-by: aszlig <aszlig@nix.build>