flake: Add jobs for multiple compilers/versions
So far we only used GCC for building the package, even on Hydra. With
our latest CLang fixes it only makes sense to actually build the package
via CLang because we want to make sure we don't introduce regressions.

Instead of just adding the default CLang compiler ("clangStdenv") I
instead added *all* of the CLang compilers starting with version 7.

I also did the same for GCC, building and testing the package with all
versions starting with version 7 as well.

All attributes for GCC and CLang will be generated up to the latest
version available in nixpkgs, so the next time we update the Flake
lockfile and there's a new compiler major version, we'll use it as well.

Signed-off-by: aszlig <aszlig@nix.build>