Bump version to 2.1.4
I'm still waiting for feedback on the systemd socket fd versus rule
number issue[1].

However since I'm pretty confident that this is fixed, even if it's not
the case, it surely did fix incrementing the rule position.

That's why I'm releasing with those fixes and if the mentioned issue is
still not fixed, we can still make a 2.1.5 release.

[1]: https://github.com/nixcloud/ip2unix/issues/27

Signed-off-by: aszlig <aszlig@nix.build>