flake: Don't override stdenv for libyamlcpp
I originally introduced this in 9fa474632160c24e43a0317494e0a92ac76ac65,
where I explicitly passed stdenv again to libyamlcpp so that we get an
appropriate clangStdenv when building with Clang.

However, libyamlcpp doesn't build with some versions of clang and since
our use of that library is already deprecated, let's build it with the
default compiler and just link it with Clang.

Signed-off-by: aszlig <aszlig@nix.build>