flake: Revert not overriding stdenv for libyamlcpp
This more or less reverts commit 2ea18f47c729d1cf5b684e7bd44b2f71678923,
where I erroneously assumed that overriding stdenv for libyamlcpp was
only because of supporting clangStdenv.

However, this is also necessary for stdenv's with different GCC versions
and linking against a library produced with a different GCC version
seems to be a lot more error-prone than with Clang.

In our case, we now get linking errors for example with
tests.full.gcc9.x86_64-linux:

  .../bin/ld: .../lib/libyaml-cpp.so: undefined reference to
  `std::__throw_bad_array_new_length()@GLIBCXX_3.4.29'

To work around this, we now no longer override stdenv when on Clang,
which should leave the GCC builds as-is (aka working) and fix the Clang
builds accordingly.

Signed-off-by: aszlig <aszlig@nix.build>