tests/rsession: Fix running "rsession" program
In newer nixpkgs versions, the "rsession" program no longer resides in
"${rstudio}/bin" but in "${rstudio}/lib/rstudio/bin" and it's expected
to be called by a wrapper.

However, in our case we only want to check if ip2unix correctly works
with the "rsession" case, so let's actually fix its path.

Signed-off-by: aszlig <aszlig@nix.build>