blackhole: Fix use-after-free of temp working dir
This is something that was pointed out by the compiler warnings of newer
versions of GCC and righfully so, because I actually declared wdir_str
which should have been returned by the function and which contains a
copy of the "workdir" variable to be freed earlier.

So freeing it at that point was actually right, but instead of returning
wdir_str, I actually returned workdir.

Fortunately, the code path with the use-after-free is only taken
whenever we were unable to find a writable temporary directory and need
to resort to writing to the current working directory.

Signed-off-by: aszlig <aszlig@nix.build>