blackhole: Document why we need BlackHole
I just was about to remove the BlackHole struct and instead use use an
unnamed socket instead.

However, when thinking about whether the unnamed socket would cause any
regressions, I remembered why I did all that tempfile stuff in the first
place.

The reason was that it provides better isolation (in terms of security),
because filesystem permissions apply and there is no connectable
abstract socket that can be simply fished out via "netstat -x".

To make sure I don't try to remove all this again in the future, I
decided to add a small comment as a reminder.

Signed-off-by: aszlig <aszlig@nix.build>