README: Separate matches/actions more clearly
So far the section on rule specification has been somewhat confusing,
since matches and actions were interleaved.

The reason why this was the case was that I was grouping them into flags
and options and probably was entrenched so deep into the technical
details that I lost sight about what actually would make more sense from
a conceptual perspective.

Signed-off-by: aszlig <aszlig@nix.build>