meson: Use .path() instead of .full_path()
This reverts commit 00fdab61fb0d0ed2489a479fde507e55f27f5563.

Using external_program.full_path() requires a minimum Meson version of
0.55.0, which was released in July 2020. While this certainly is old
enough for the distro I'm using, it might be too new for other distros
that can't easily mix different versions and/or are on long-term
support.

Additionally, I was confused by all the .full_path() methods I used in
the same file, but it turned out they weren't for the external_program
object.

This confusion is probably the reason why Meson changed this to be more
consistent in version 0.55.0.

Signed-off-by: aszlig <aszlig@nix.build>