sockaddr: Wrap instead of inheriting inner type
So far we just directly subclassed sockaddr_storage and all operations
were just casting around raw memory.

This might sound like a good idea at first since after all we just want
to add a bunch of helper methods but without changing much about how the
data is stored.

It also works well for all the socket types we had so far, which
essentially are sockaddr_in, sockaddr_in6 and sockaddr_un. This is
because all of them *seem* to have a static size that never relies on
the actual content.

Unless it even doesn't work for sockaddr_un. Quoting from unix(7):

  An abstract socket address is distinguished (from a pathname socket)
  by the fact that sun_path[0] is a null byte ('\0'). The socket's
  address in this namespace is given by the additional bytes in sun_path
  that are covered by the specified length of the address structure.
  (Null bytes in the name have no special significance.) The name has no
  connection with filesystem pathnames. When the address of an abstract
  socket is returned, the returned addrlen is greater than
  sizeof(sa_family_t) (i.e., greater than 2), and the name of the socket
  is contained in the first (addrlen - sizeof(sa_family_t)) bytes of
  sun_path.

This means that we no longer can calculate the size on the fly but
instead store it as a private field in our SockAddr wrapper.

The reason I decided to do this was when implementing support for
abstract sockets, but since this is something I generally consider an
improvement, I decided to apply this change to the master branch.

Signed-off-by: aszlig <aszlig@nix.build>