Rename dynports.cc to dynports.cc.in
The file in question is processed during the build process and the
actual source file is the generated output of genoffsets.py.

It's somewhat confusing if the input file has the same .cc ending, since
this would hint that it's actually a valid C++ file, which is not.

When adding clang-tidy to the mix of tooling we're using for static
analysis, it was tripping over this file. I could have excluded the
src/dynports directory, but we would manage those exclusions for every
single static analyser.

Instead, I renamed the file and gave it an .in suffix, which is very
common among various build systems (eg. autotools, CMake) and thus will
also be less of a surprise for developers not familiar to the project.

Signed-off-by: aszlig <aszlig@nix.build>