Selectively disable some Clang-Tidy lints
I've disabled the performance-unnecessary-value-param lint for the
shared_ptr passed to Socket::when closure, because the lint would
suggest to use something like "const Socket::Ptr &", which is
essentially "const std::shared_ptr<Socket> &". Since this already
implies a pointer, it doesn't make sense to wrap it behind yet another
reference.

A few of the bugprone-macro-parentheses lints were either suggesting
putting parenthesis around something that would result in a syntax error
or would actually change the behaviour since not using parenthesis was
intentional (I know, I know...).

The bugprone-exception-escape lints were only showing up for the main
functions of our unit tests, so it's not really a real issue as long as
the tests are working as intended.

Signed-off-by: aszlig <aszlig@nix.build>