preload: Initialise ssize_t instead of casting
Aside from being unnecessary, a cast also has the negative side-effect
that it will cause a warning on certain architectures (eg. i686-linux)
where ssize_t is equal to an int.

I'm personally not a big fan of brace-initialisation since it looks
ugly, but since ssize_t literals are only supported in C++23 and we're
currently targetting C++17, we're stuck with brace-initialisation as the
lesser evil.

Signed-off-by: aszlig <aszlig@nix.build>