flake: Move clang-tidy jobs into checks
The checks Flake attribute contains derivations that are built whenever
we perform a "nix flake check" and given that Clang-Tidy linting our
code is only really useful if we make sure that we never push code that
has Clang-Tidy failures.

To enforce this, let's move the clang-tidy Hydra job into the checks
attribute, that way we can easily run it prior to pushing and keep our
repository free from linting errors.

Of course, I also made sure that the clang-tidy job is built on Hydra as
well, by reintroducing the attribute under the "basic" job namespace.

Signed-off-by: aszlig <aszlig@nix.build>