github: Prevent Actions running twice in PRs
This is seems to be a common[1] issue[2] and there really is not a very
satisfactory solution other than adding an "if" on the job rather than
on the top-level workflow. Of course, there were other solutions listed,
but they'd involve managing a hardcoded list of branch names where the
"on push" action is triggered.

So I went for the workaround listed in [1], which adds an if to the job
specification.

[1]: https://github.com/zopefoundation/meta/issues/145
[2]: https://github.com/open-telemetry/opentelemetry-python/issues/1370

Signed-off-by: aszlig <aszlig@nix.build>