Require at least GCC C++ compiler version 9
So far, our "oldest supported distro" was Ubuntu 18.04 LTS, but since
that version no longer gets standard support from Canonical[1], we can
safely assume that apart from a few enterprise users noone is using it
anymore. Even in case there are any enterprise ip2unix users, they can
still use an older version of ip2unix.

The reason why we want to get rid of GCC C++ version 7 is because the
C++17 Filesystem library wasn't implemented back then and only got an
incomplete implementation starting with GCC version 9.

However, we really want to use the Filesystem library because it allows
us to work with files in a cross-platform way, which it turn makes it
easier to port ip2unix to eg. Darwin/BSD.

I also changed the GitHub workflow to use ubuntu-20.04 instead of
ubuntu-latest because our actual tests run on Hydra and we're only using
GitHub Actions to make sure it works in the worlds outside of Nix. Using
the oldest supported LTS instead of the latest non-LTS version of Ubuntu
makes sure we're not accidentally breaking things for older software
since with nixpkgs we usually use recent software and thus already cover
this with our Hydra builds.

[1]: https://ubuntu.com/18-04

Signed-off-by: aszlig <aszlig@nix.build>