tests/abstract: Use UUID for the abstract name
The repeat100 Hydra jobsets fail regularly with errors like these:

      @abstract_sockets_only
      def test_abstcact_socket():
          with socket.socket(socket.AF_UNIX, socket.SOCK_STREAM) as sock:
  >           sock.bind('\0abstest')
  E           OSError: [Errno 98] Address already in use

This of course is because we always use the same abstract socket name,
which at some point might not be cleaned up before the next test run
starts, resulting in the above error.

Simply using an UUID as the name should avoid this.

Signed-off-by: aszlig <aszlig@nix.build>