tests: Synchronise when running abstract connector
By default, the run_server method doesn't wait until the socket is ready
and hands over to the context manager as soon as the connector process
is started.

This obviously is prone to race conditions where we get errors whenever
the client is trying to connect too early.

We do have a sync keyword argument for this already, so all we need to
do to make the test more stable is to enable it.

I also added a docstring which explains the keyword arguments a bit so
that next time I hopefully won't forget about it.

Signed-off-by: aszlig <aszlig@nix.build>