tests/connector: Switch to UUID for abstract name
This is similar to 2833273873d6a5e134618acd9cb37e913e7bd90d, where we
used UUIDs for the dedicated abstract test. However, when running the
repeat100 tests, we also ran into collisions with the connector, so
let's actually switch the connector to use UUIDs as well to avoid this.

Signed-off-by: aszlig <aszlig@nix.build>