Internally use proper sotypes instead of TCP/UDP
So far we've exclusively focussed on TCP and UDP as the socket type, but
technically the correct term would be stream and datagram socket types.

At some point we're going to have the ability to work on *existing* Unix
domain sockets and using TCP and UDP as names for the socket type is
really confusing, since it's not actually a real UDP/TCP socket.

We still allow to use "udp" and "tcp" for rule definitions and also use
"UDP" and "TCP" when displaying the parsed rules, but we no longer use
it internally.

Signed-off-by: aszlig <aszlig@nix.build>