socket: Rename is_unix to is_converted
While "is_unix" certainly would describe the state after being converted
from an AF_INET/AF_INET6 socket to an AF_UNIX socket, it's not
necessarily true if we pick up an existing AF_UNIX socket. In this case
the "conversion" would be from AF_UNIX to AF_UNIX and the name "is_unix"
would be ambiguous and could apply to the state before the conversion
*and* afterwards.

I stumbled upon this when implementing a way to pick up an existing Unix
domain socket and eg. change its path.

Signed-off-by: aszlig <aszlig@nix.build>