rules: Split Rule struct into matches and action
In 242c78d68b46282983e7658e67083f9b96fc086f, we improved the README file
by more clearly distinguishing between matches and an action of a rule.

However, the code did not yet reflect this and all rule arguments were
thrown together into one single struct without this important
distinction.

It makes sense to actually model our data structures in the same terms
we use in our documentation, so this is only a start of that.

The reason why I'm using the plural "matches" is because we can have
several of them, while onle one action is possible.

Right now, the Rule.action field is still a struct, but ideally this
should be an std::variant so that we can enforce this constraint at the
type level.

Since std::variant and particularly std::visit is very gruesome compared
to eg. the "match" keyword in Rust, "case" in Haskell or any other
language with proper algebraic data types, moving to std::variant will
be more involved. For now however, splitting Rule into these two
sub-structs is at least a move in the right direction.

Signed-off-by: aszlig <aszlig@nix.build>