Revert "Use dlopen on libc instead of RTLD_NEXT"
This reverts commit f47125c0513ecc484c8f537e9f496d601c3ec7b4.

The commit in question directly used the C library instead of RTLD_NEXT
whenever possible, based on some deadlock I didn't specify further in
the commit message.

I could not reproduce a deadlock like this and given that we don't have
a test case that fails with RTLD_NEXT tells me that it doesn't matter.

In addition to just reverting, I also added a small test case that I
used in order to try reproducing the aforementioned issue. Even though I
couldn't reproduce the issue, I'm adding the test anyway since we didn't
have a test case where we use multiple LD_PRELOAD wrappers that operate
on the same set of calls as ip2unix.

Signed-off-by: aszlig <aszlig@nix.build>