socketpath: Add missing std::hash implementation
When adding support for abstract sockets in commit
a1b0a51ec4a337c27001e11cdc9a3ac4b7cebc8c, I accidentally forgot to add
the actual hash implementation for the new SocketPath struct, which I
marked as TODO but then forgot.

The practical implications of this are really neglicible since most of
the time users won't mix abstract and filesystem sockets with the same
name in gazillion of rules so that hash collisions would make a
performance impact.

However, the *real* impact is that the "// TODO" comment is a major
annoyance to me, so with this commit it's gone :-)

Signed-off-by: aszlig <aszlig@nix.build>