Use Meson option for abstract socket support
So far, we checked whether the kernel is Linux, which generally is a
good indicator since abstract sockets are only supported on Linux.

The disadvantage of checking for defined(__linux__) is that we can't
simply disable it, except if we'd switch kernels. Disabling support
however is necessary so that we can check whether compilation still
succeeeds with that feature disabled.

Signed-off-by: aszlig <aszlig@nix.build>