INSTALL: Fix broken link to systemd manpage
The manpage link for the systemd-socket-activate tool was using an
attribute (sysdman_url) as its base-URL, which we only defined in
README.adoc. Since splitting the installation instructions off the main
README in 990707df616f34248c699420b140137373ae62a7, that attribute is no
longer defined and the link gets printed raw.

Signed-off-by: aszlig <aszlig@nix.build>