Bump version to 2.2.0
This release already is quite packed with features and most of the
things I wanted to get in (eg. abstract sockets and the from-* options)
are done.

Given that I'm moderately happy with the current state and also the fact
that we now emit a deprecation warning for YAML, it's just more
important that this warning is actually displayed to potential users.

Signed-off-by: aszlig <aszlig@nix.build>