preload: Add logging for rule matching
I was debugging something that turned out to be a serialisation issue
and was wondering why a rule didn't match properly.

In the end, I found out the culprit via debugger, but those log messages
are helpful not even for developers but also users to check why a
particular rule doesn't or does match.

Signed-off-by: aszlig <aszlig@nix.build>