Fix missing serialisation for from-{unix,abstract}
When adding support for transforming paths/names from existing Unix
domain sockets in e332bf5f61759d847f6f58d5cfe6d87b1b31973b, I forgot to
actually add the new fields to the serialisation methods.

The effect of this is that when using from-*, we suddenly match *all*
the sockets, regardless of whether it's an existing Unix domain socket
or not.

Our test case didn't catch that because it ran in isolation with only
one rule, so I modified the test case to add an unrelated socket that
should not be picked up by ip2unix.

Signed-off-by: aszlig <aszlig@nix.build>